third_party/blink/web_tests/wpt_internal/clipboard-apis/async-custom-format-write-read.tentative.https.html

<!doctype html>
<meta charset="utf-8">
<title>Async Clipboard Custom Format platform-specific write -> read tests</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<body>Body needed for test_driver.click()</body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/user-activation.js"></script>
<script>
'use strict';

async function customFormatWriteBeforeTest() {
  const format = 'web text/customformat';
  const blob = new Blob(['CustomFormatClipboardDisabled'], {type: format});
  const clipboardItem = new ClipboardItem({[format]: blob});
  await waitForUserActivation();
  await navigator.clipboard.write([clipboardItem]);
}

// Writes a payload with custom content and checks to ensure the correct data
// was written successfully.
promise_test(async t => {
  // TODO(crbug.com/106449): Enable after implementing custom formats read/write.
  await test_driver.set_permission({name: 'clipboard-write', allowWithoutSanitization: true}, 'granted');
  await test_driver.set_permission({name: 'clipboard-write', allowWithoutSanitization: false},
                                        'granted');
  // This extra custom format write is used to create consistency in the error
  // message when the ClipboardCustomFormats flag isn't enabled.
  await customFormatWriteBeforeTest();
  const dataToWrite = 'Test text.';
  const format = 'web text/plain';

  const blobInput = new Blob([dataToWrite], {type: format});
  // Blob types are automatically converted to lower-case.
  assert_equals(blobInput.type, format.toLowerCase());
  const clipboardItemInput = new ClipboardItem({[format]: blobInput});
  await waitForUserActivation();
  await navigator.clipboard.write([clipboardItemInput]);

  // Items should be readable on a system clipboard after custom format write.
  await waitForUserActivation();
  const clipboardItems = await navigator.clipboard.read();
  assert_equals(clipboardItems.length, 1);
  const clipboardItem = clipboardItems[0];
  assert_true(clipboardItem instanceof ClipboardItem);

  const blobOutput = await clipboardItem.getType(format);
  assert_equals(blobOutput.type, format);
  const data = await (new Response(blobOutput)).text();
  assert_equals(data, dataToWrite);

  // These examples use web custom text format, so this format shouldn't be
  // accessible as text.
  const textOutput = await navigator.clipboard.readText();
  assert_not_equals(textOutput, dataToWrite);
}, 'Verify write and read clipboard given web custom format as input');
</script>
