third_party/blink/web_tests/platform/linux-chrome/wpt_internal/clipboard-apis/async-custom-format-write-read.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] Verify write and read clipboard given web custom format as input
  assert_equals: expected "Test text." but got ""
Harness: the test ran to completion.
